Move native_assets_tests into isolated/
