+++ packages/flutter_tools/test/general.shard/isolated/build_system/targets/native_assets_test.dart
@@ -17,9 +17,9 @@ import 'package:native_assets_cli/native_assets_cli_internal.dart'
     as native_assets_cli;
 import 'package:package_config/package_config.dart' show Package;
 
-import '../../../src/common.dart';
-import '../../../src/context.dart';
-import '../../../src/fakes.dart';
+import '../../../../src/common.dart';
+import '../../../../src/context.dart';
+import '../../../../src/fakes.dart';
 import '../../fake_native_assets_build_runner.dart';
 
 void main() {
@@ -66,9 +66,11 @@ void main() {
     androidEnvironment.buildDir.createSync(recursive: true);
   });
 
-  testWithoutContext('NativeAssets throws error if missing target platform', () async {
+  testWithoutContext('NativeAssets throws error if missing target platform',
+      () async {
     iosEnvironment.defines.remove(kTargetPlatform);
-    expect(const NativeAssets().build(iosEnvironment), throwsA(isA<MissingDefineException>()));
+    expect(const NativeAssets().build(iosEnvironment),
+        throwsA(isA<MissingDefineException>()));
   });
 
   testUsingContext('NativeAssets defaults to ios archs if missing', () async {
@@ -90,7 +92,8 @@ void main() {
     await createPackageConfig(iosEnvironment);
 
     iosEnvironment.defines.remove(kSdkRoot);
-    expect(const NativeAssets().build(iosEnvironment), throwsA(isA<MissingDefineException>()));
+    expect(const NativeAssets().build(iosEnvironment),
+        throwsA(isA<MissingDefineException>()));
   });
 
   // The NativeAssets Target should _always_ be creating a yaml an d file.
@@ -103,13 +106,14 @@ void main() {
         FileSystem: () => fileSystem,
         ProcessManager: () => processManager,
         FeatureFlags: () => TestFeatureFlags(
-          isNativeAssetsEnabled: isNativeAssetsEnabled,
-        ),
+              isNativeAssetsEnabled: isNativeAssetsEnabled,
+            ),
       },
       () async {
         await createPackageConfig(iosEnvironment);
 
-        final NativeAssetsBuildRunner buildRunner = FakeNativeAssetsBuildRunner();
+        final NativeAssetsBuildRunner buildRunner =
+            FakeNativeAssetsBuildRunner();
         await NativeAssets(buildRunner: buildRunner).build(iosEnvironment);
 
         expect(iosEnvironment.buildDir.childFile('native_assets.d'), exists);
@@ -129,8 +133,11 @@ void main() {
       await createPackageConfig(iosEnvironment);
 
       final NativeAssetsBuildRunner buildRunner = FakeNativeAssetsBuildRunner(
-        packagesWithNativeAssetsResult: <Package>[Package('foo', iosEnvironment.buildDir.uri)],
-        buildResult: FakeNativeAssetsBuilderResult(assets: <native_assets_cli.Asset>[
+        packagesWithNativeAssetsResult: <Package>[
+          Package('foo', iosEnvironment.buildDir.uri)
+        ],
+        buildResult:
+            FakeNativeAssetsBuilderResult(assets: <native_assets_cli.Asset>[
           native_assets_cli.Asset(
             id: 'package:foo/foo.dart',
             linkMode: native_assets_cli.LinkMode.dynamic,
@@ -145,8 +152,10 @@ void main() {
       );
       await NativeAssets(buildRunner: buildRunner).build(iosEnvironment);
 
-      final File nativeAssetsYaml = iosEnvironment.buildDir.childFile('native_assets.yaml');
-      final File depsFile = iosEnvironment.buildDir.childFile('native_assets.d');
+      final File nativeAssetsYaml =
+          iosEnvironment.buildDir.childFile('native_assets.yaml');
+      final File depsFile =
+          iosEnvironment.buildDir.childFile('native_assets.d');
       expect(depsFile, exists);
       // We don't care about the specific format, but it should contain the
       // yaml as the file depending on the source files that went in to the
