+++ packages/flutter_tools/test/general.shard/isolated/macos/native_assets_host_test.dart
@@ -4,7 +4,7 @@
 
 import 'package:flutter_tools/src/isolated/native_assets/macos/native_assets_host.dart';
 
-import '../../src/common.dart';
+import '../../../src/common.dart';
 
 void main() {
   test('framework name', () {
