+++ packages/flutter_tools/test/general.shard/isolated/windows/native_assets_test.dart
@@ -6,7 +6,6 @@ import 'package:file/file.dart';
 import 'package:file/memory.dart';
 import 'package:file_testing/file_testing.dart';
 import 'package:flutter_tools/src/artifacts.dart';
-import 'package:flutter_tools/src/base/common.dart';
 import 'package:flutter_tools/src/base/file_system.dart';
 import 'package:flutter_tools/src/base/logger.dart';
 import 'package:flutter_tools/src/base/platform.dart';
@@ -15,17 +14,17 @@ import 'package:flutter_tools/src/build_system/build_system.dart';
 import 'package:flutter_tools/src/dart/package_map.dart';
 import 'package:flutter_tools/src/features.dart';
 import 'package:flutter_tools/src/globals.dart' as globals;
-import 'package:flutter_tools/src/isolated/native_assets/linux/native_assets.dart';
 import 'package:flutter_tools/src/isolated/native_assets/native_assets.dart';
+import 'package:flutter_tools/src/isolated/native_assets/windows/native_assets.dart';
 import 'package:native_assets_cli/native_assets_cli_internal.dart'
     hide BuildMode, Target;
 import 'package:native_assets_cli/native_assets_cli_internal.dart'
     as native_assets_cli;
 import 'package:package_config/package_config_types.dart';
 
-import '../../src/common.dart';
-import '../../src/context.dart';
-import '../../src/fakes.dart';
+import '../../../src/common.dart';
+import '../../../src/context.dart';
+import '../../../src/fakes.dart';
 import '../fake_native_assets_build_runner.dart';
 
 void main() {
@@ -40,7 +39,7 @@ void main() {
     processManager = FakeProcessManager.empty();
     logger = BufferLogger.test();
     artifacts = Artifacts.test();
-    fileSystem = MemoryFileSystem.test();
+    fileSystem = MemoryFileSystem.test(style: FileSystemStyle.windows);
     environment = Environment.test(
       fileSystem.currentDirectory,
       inputs: <String, String>{},
@@ -53,11 +52,12 @@ void main() {
     projectUri = environment.projectDir.uri;
   });
 
-  testUsingContext('dry run with no package config', overrides: <Type, Generator>{
-    ProcessManager: () => FakeProcessManager.empty(),
-  }, () async {
+  testUsingContext('dry run with no package config',
+      overrides: <Type, Generator>{
+        ProcessManager: () => FakeProcessManager.empty(),
+      }, () async {
     expect(
-      await dryRunNativeAssetsLinux(
+      await dryRunNativeAssetsWindows(
         projectUri: projectUri,
         fileSystem: fileSystem,
         buildRunner: FakeNativeAssetsBuildRunner(
@@ -75,7 +75,7 @@ void main() {
   testUsingContext('build with no package config', overrides: <Type, Generator>{
     ProcessManager: () => FakeProcessManager.empty(),
   }, () async {
-    await buildNativeAssetsLinux(
+    await buildNativeAssetsWindows(
       projectUri: projectUri,
       buildMode: BuildMode.debug,
       fileSystem: fileSystem,
@@ -89,33 +89,15 @@ void main() {
     );
   });
 
-  testUsingContext('does not throw if clang not present but no native assets present', overrides: <Type, Generator>{
-    FeatureFlags: () => TestFeatureFlags(isNativeAssetsEnabled: true),
-    ProcessManager: () => FakeProcessManager.empty(),
-  }, () async {
-    final File packageConfig = environment.projectDir.childFile('.dart_tool/package_config.json');
-    await packageConfig.create(recursive: true);
-    await buildNativeAssetsLinux(
-      projectUri: projectUri,
-      buildMode: BuildMode.debug,
-      fileSystem: fileSystem,
-      buildRunner: _BuildRunnerWithoutClang(),
-    );
-    expect(
-      (globals.logger as BufferLogger).traceText,
-      isNot(contains('Building native assets for ')),
-    );
-  });
-
-  testUsingContext('dry run for multiple OSes with no package config', overrides: <Type, Generator>{
-    ProcessManager: () => FakeProcessManager.empty(),
-  }, () async {
+  testUsingContext('dry run for multiple OSes with no package config',
+      overrides: <Type, Generator>{
+        ProcessManager: () => FakeProcessManager.empty(),
+      }, () async {
     await dryRunNativeAssetsMultipleOSes(
       projectUri: projectUri,
       fileSystem: fileSystem,
       targetPlatforms: <TargetPlatform>[
-        TargetPlatform.darwin,
-        TargetPlatform.ios,
+        TargetPlatform.windows_x64,
       ],
       buildRunner: FakeNativeAssetsBuildRunner(
         hasPackageConfigResult: false,
@@ -127,14 +109,16 @@ void main() {
     );
   });
 
-  testUsingContext('dry run with assets but not enabled', overrides: <Type, Generator>{
-    ProcessManager: () => FakeProcessManager.empty(),
-  }, () async {
-    final File packageConfig = environment.projectDir.childFile('.dart_tool/package_config.json');
+  testUsingContext('dry run with assets but not enabled',
+      overrides: <Type, Generator>{
+        ProcessManager: () => FakeProcessManager.empty(),
+      }, () async {
+    final File packageConfig =
+        environment.projectDir.childFile('.dart_tool/package_config.json');
     await packageConfig.parent.create();
     await packageConfig.create();
     expect(
-      () => dryRunNativeAssetsLinux(
+      () => dryRunNativeAssetsWindows(
         projectUri: projectUri,
         fileSystem: fileSystem,
         buildRunner: FakeNativeAssetsBuildRunner(
@@ -144,7 +128,8 @@ void main() {
         ),
       ),
       throwsToolExit(
-        message: 'Package(s) bar require the native assets feature to be enabled. '
+        message:
+            'Package(s) bar require the native assets feature to be enabled. '
             'Enable using `flutter config --enable-native-assets`.',
       ),
     );
@@ -154,10 +139,11 @@ void main() {
     FeatureFlags: () => TestFeatureFlags(isNativeAssetsEnabled: true),
     ProcessManager: () => FakeProcessManager.empty(),
   }, () async {
-    final File packageConfig = environment.projectDir.childFile('.dart_tool/package_config.json');
+    final File packageConfig =
+        environment.projectDir.childFile('.dart_tool/package_config.json');
     await packageConfig.parent.create();
     await packageConfig.create();
-    final Uri? nativeAssetsYaml = await dryRunNativeAssetsLinux(
+    final Uri? nativeAssetsYaml = await dryRunNativeAssetsWindows(
       projectUri: projectUri,
       fileSystem: fileSystem,
       buildRunner: FakeNativeAssetsBuildRunner(
@@ -169,14 +155,8 @@ void main() {
             Asset(
               id: 'package:bar/bar.dart',
               linkMode: LinkMode.dynamic,
-              target: native_assets_cli.Target.linuxX64,
-              path: AssetAbsolutePath(Uri.file('libbar.so')),
-            ),
-            Asset(
-              id: 'package:bar/bar.dart',
-              linkMode: LinkMode.dynamic,
-              target: native_assets_cli.Target.linuxArm64,
-              path: AssetAbsolutePath(Uri.file('libbar.so')),
+              target: native_assets_cli.Target.windowsX64,
+              path: AssetAbsolutePath(Uri.file('bar.dll')),
             ),
           ],
         ),
@@ -185,13 +165,13 @@ void main() {
     expect(
       (globals.logger as BufferLogger).traceText,
       stringContainsInOrder(<String>[
-        'Dry running native assets for linux.',
-        'Dry running native assets for linux done.',
+        'Dry running native assets for windows.',
+        'Dry running native assets for windows done.',
       ]),
     );
     expect(
       nativeAssetsYaml,
-      projectUri.resolve('build/native_assets/linux/native_assets.yaml'),
+      projectUri.resolve('build/native_assets/windows/native_assets.yaml'),
     );
     expect(
       await fileSystem.file(nativeAssetsYaml).readAsString(),
@@ -199,14 +179,16 @@ void main() {
     );
   });
 
-  testUsingContext('build with assets but not enabled', overrides: <Type, Generator>{
-    ProcessManager: () => FakeProcessManager.empty(),
-  }, () async {
-    final File packageConfig = environment.projectDir.childFile('.dart_tool/package_config.json');
+  testUsingContext('build with assets but not enabled',
+      overrides: <Type, Generator>{
+        ProcessManager: () => FakeProcessManager.empty(),
+      }, () async {
+    final File packageConfig =
+        environment.projectDir.childFile('.dart_tool/package_config.json');
     await packageConfig.parent.create();
     await packageConfig.create();
     expect(
-      () => buildNativeAssetsLinux(
+      () => buildNativeAssetsWindows(
         projectUri: projectUri,
         buildMode: BuildMode.debug,
         fileSystem: fileSystem,
@@ -217,7 +199,8 @@ void main() {
         ),
       ),
       throwsToolExit(
-        message: 'Package(s) bar require the native assets feature to be enabled. '
+        message:
+            'Package(s) bar require the native assets feature to be enabled. '
             'Enable using `flutter config --enable-native-assets`.',
       ),
     );
@@ -227,11 +210,12 @@ void main() {
     FeatureFlags: () => TestFeatureFlags(isNativeAssetsEnabled: true),
     ProcessManager: () => FakeProcessManager.empty(),
   }, () async {
-    final File packageConfig = environment.projectDir.childFile('.dart_tool/package_config.json');
+    final File packageConfig =
+        environment.projectDir.childFile('.dart_tool/package_config.json');
     await packageConfig.parent.create();
     await packageConfig.create();
-    final (Uri? nativeAssetsYaml, _) = await buildNativeAssetsLinux(
-      targetPlatform: TargetPlatform.linux_x64,
+    final (Uri? nativeAssetsYaml, _) = await buildNativeAssetsWindows(
+      targetPlatform: TargetPlatform.windows_x64,
       projectUri: projectUri,
       buildMode: BuildMode.debug,
       fileSystem: fileSystem,
@@ -243,7 +227,7 @@ void main() {
     );
     expect(
       nativeAssetsYaml,
-      projectUri.resolve('build/native_assets/linux/native_assets.yaml'),
+      projectUri.resolve('build/native_assets/windows/native_assets.yaml'),
     );
     expect(
       await fileSystem.file(nativeAssetsYaml).readAsString(),
@@ -253,7 +237,7 @@ void main() {
       environment.projectDir
           .childDirectory('build')
           .childDirectory('native_assets')
-          .childDirectory('linux'),
+          .childDirectory('windows'),
       exists,
     );
   });
@@ -267,14 +251,16 @@ void main() {
       FeatureFlags: () => TestFeatureFlags(isNativeAssetsEnabled: true),
       ProcessManager: () => FakeProcessManager.empty(),
     }, () async {
-      final File packageConfig = environment.projectDir.childDirectory('.dart_tool').childFile('package_config.json');
+      final File packageConfig = environment.projectDir
+          .childDirectory('.dart_tool')
+          .childFile('package_config.json');
       await packageConfig.parent.create();
       await packageConfig.create();
-      final File dylibAfterCompiling = fileSystem.file('libbar.so');
+      final File dylibAfterCompiling = fileSystem.file('bar.dll');
       // The mock doesn't create the file, so create it here.
       await dylibAfterCompiling.create();
-      final (Uri? nativeAssetsYaml, _) = await buildNativeAssetsLinux(
-        targetPlatform: TargetPlatform.linux_x64,
+      final (Uri? nativeAssetsYaml, _) = await buildNativeAssetsWindows(
+        targetPlatform: TargetPlatform.windows_x64,
         projectUri: projectUri,
         buildMode: BuildMode.debug,
         fileSystem: fileSystem,
@@ -288,7 +274,7 @@ void main() {
               Asset(
                 id: 'package:bar/bar.dart',
                 linkMode: LinkMode.dynamic,
-                target: native_assets_cli.Target.linuxX64,
+                target: native_assets_cli.Target.windowsX64,
                 path: AssetAbsolutePath(dylibAfterCompiling.uri),
               ),
             ],
@@ -298,13 +284,13 @@ void main() {
       expect(
         (globals.logger as BufferLogger).traceText,
         stringContainsInOrder(<String>[
-          'Building native assets for linux_x64 debug.',
-          'Building native assets for linux_x64 done.',
+          'Building native assets for windows_x64 debug.',
+          'Building native assets for windows_x64 done.',
         ]),
       );
       expect(
         nativeAssetsYaml,
-        projectUri.resolve('build/native_assets/linux/native_assets.yaml'),
+        projectUri.resolve('build/native_assets/windows/native_assets.yaml'),
       );
       expect(
         await fileSystem.file(nativeAssetsYaml).readAsString(),
@@ -312,10 +298,10 @@ void main() {
           'package:bar/bar.dart',
           if (flutterTester)
             // Tests run on host system, so the have the full path on the system.
-            '- ${projectUri.resolve('build/native_assets/linux/libbar.so').toFilePath()}'
+            '- ${projectUri.resolve('build/native_assets/windows/bar.dll').toFilePath()}'
           else
             // Apps are a bundle with the dylibs on their dlopen path.
-            '- libbar.so',
+            '- bar.dll',
         ]),
       );
     });
@@ -325,11 +311,12 @@ void main() {
     FeatureFlags: () => TestFeatureFlags(isNativeAssetsEnabled: true),
     ProcessManager: () => FakeProcessManager.empty(),
   }, () async {
-    final File packageConfig = environment.projectDir.childFile('.dart_tool/package_config.json');
+    final File packageConfig =
+        environment.projectDir.childFile('.dart_tool/package_config.json');
     await packageConfig.parent.create();
     await packageConfig.create();
     expect(
-      () => dryRunNativeAssetsLinux(
+      () => dryRunNativeAssetsWindows(
         projectUri: projectUri,
         fileSystem: fileSystem,
         buildRunner: FakeNativeAssetsBuildRunner(
@@ -341,21 +328,17 @@ void main() {
               Asset(
                 id: 'package:bar/bar.dart',
                 linkMode: LinkMode.static,
-                target: native_assets_cli.Target.macOSArm64,
-                path: AssetAbsolutePath(Uri.file('bar.a')),
-              ),
-              Asset(
-                id: 'package:bar/bar.dart',
-                linkMode: LinkMode.static,
-                target: native_assets_cli.Target.macOSX64,
-                path: AssetAbsolutePath(Uri.file('bar.a')),
+                target: native_assets_cli.Target.windowsX64,
+                path: AssetAbsolutePath(
+                    Uri.file(OS.windows.staticlibFileName('bar'))),
               ),
             ],
           ),
         ),
       ),
       throwsToolExit(
-        message: 'Native asset(s) package:bar/bar.dart have their link mode set to '
+        message:
+            'Native asset(s) package:bar/bar.dart have their link mode set to '
             'static, but this is not yet supported. '
             'For more info see https://github.com/dart-lang/sdk/issues/49418.',
       ),
@@ -371,7 +354,7 @@ void main() {
     await packageConfig.parent.create();
     await packageConfig.create();
     expect(
-      () => dryRunNativeAssetsLinux(
+      () => dryRunNativeAssetsWindows(
         projectUri: projectUri,
         fileSystem: fileSystem,
         buildRunner: FakeNativeAssetsBuildRunner(
@@ -399,8 +382,8 @@ void main() {
     await packageConfig.parent.create();
     await packageConfig.create();
     expect(
-      () => buildNativeAssetsLinux(
-        targetPlatform: TargetPlatform.linux_x64,
+      () => buildNativeAssetsWindows(
+        targetPlatform: TargetPlatform.windows_x64,
         projectUri: projectUri,
         buildMode: BuildMode.debug,
         fileSystem: fileSystem,
@@ -424,29 +407,91 @@ void main() {
   // This logic is mocked in the other tests to avoid having test order
   // randomization causing issues with what processes are invoked.
   // Exercise the parsing of the process output in this separate test.
-  testUsingContext('NativeAssetsBuildRunnerImpl.cCompilerConfig', overrides: <Type, Generator>{
-    FeatureFlags: () => TestFeatureFlags(isNativeAssetsEnabled: true),
-    ProcessManager: () => FakeProcessManager.list(
-          <FakeCommand>[
-            const FakeCommand(
-              command: <Pattern>['which', 'clang++'],
-              stdout: '''
-/some/path/to/clang++
+  testUsingContext('NativeAssetsBuildRunnerImpl.cCompilerConfig',
+      overrides: <Type, Generator>{
+        FeatureFlags: () => TestFeatureFlags(isNativeAssetsEnabled: true),
+        ProcessManager: () => FakeProcessManager.list(
+              <FakeCommand>[
+                FakeCommand(
+                  command: <Pattern>[
+                    RegExp(r'(.*)vswhere.exe'),
+                    '-format',
+                    'json',
+                    '-products',
+                    '*',
+                    '-utf8',
+                    '-latest',
+                    '-version',
+                    '16',
+                    '-requires',
+                    'Microsoft.VisualStudio.Workload.NativeDesktop',
+                    'Microsoft.VisualStudio.Component.VC.Tools.x86.x64',
+                    'Microsoft.VisualStudio.Component.VC.CMake.Project',
+                  ],
+                  stdout: r'''
+[
+  {
+    "instanceId": "491ec752",
+    "installDate": "2023-04-21T08:17:11Z",
+    "installationName": "VisualStudio/17.5.4+33530.505",
+    "installationPath": "C:\\Program Files\\Microsoft Visual Studio\\2022\\Community",
+    "installationVersion": "17.5.33530.505",
+    "productId": "Microsoft.VisualStudio.Product.Community",
+    "productPath": "C:\\Program Files\\Microsoft Visual Studio\\2022\\Community\\Common7\\IDE\\devenv.exe",
+    "state": 4294967295,
+    "isComplete": true,
+    "isLaunchable": true,
+    "isPrerelease": false,
+    "isRebootRequired": false,
+    "displayName": "Visual Studio Community 2022",
+    "description": "Powerful IDE, free for students, open-source contributors, and individuals",
+    "channelId": "VisualStudio.17.Release",
+    "channelUri": "https://aka.ms/vs/17/release/channel",
+    "enginePath": "C:\\Program Files (x86)\\Microsoft Visual Studio\\Installer\\resources\\app\\ServiceHub\\Services\\Microsoft.VisualStudio.Setup.Service",
+    "installedChannelId": "VisualStudio.17.Release",
+    "installedChannelUri": "https://aka.ms/vs/17/release/channel",
+    "releaseNotes": "https://docs.microsoft.com/en-us/visualstudio/releases/2022/release-notes-v17.5#17.5.4",
+    "thirdPartyNotices": "https://go.microsoft.com/fwlink/?LinkId=661288",
+    "updateDate": "2023-04-21T08:17:11.2249473Z",
+    "catalog": {
+      "buildBranch": "d17.5",
+      "buildVersion": "17.5.33530.505",
+      "id": "VisualStudio/17.5.4+33530.505",
+      "localBuild": "build-lab",
+      "manifestName": "VisualStudio",
+      "manifestType": "installer",
+      "productDisplayVersion": "17.5.4",
+      "productLine": "Dev17",
+      "productLineVersion": "2022",
+      "productMilestone": "RTW",
+      "productMilestoneIsPreRelease": "False",
+      "productName": "Visual Studio",
+      "productPatchVersion": "4",
+      "productPreReleaseMilestoneSuffix": "1.0",
+      "productSemanticVersion": "17.5.4+33530.505",
+      "requiredEngineVersion": "3.5.2150.18781"
+    },
+    "properties": {
+      "campaignId": "2060:abb99c5d1ecc4013acf2e1814b10b690",
+      "channelManifestId": "VisualStudio.17.Release/17.5.4+33530.505",
+      "nickname": "",
+      "setupEngineFilePath": "C:\\Program Files (x86)\\Microsoft Visual Studio\\Installer\\setup.exe"
+    }
+  }
+]
 ''', // Newline at the end of the string.
-            )
-          ],
-        ),
-    FileSystem: () => fileSystem,
-  }, () async {
-    if (!const LocalPlatform().isLinux) {
+                )
+              ],
+            ),
+        FileSystem: () => fileSystem,
+      }, () async {
+    if (!const LocalPlatform().isWindows) {
       return;
     }
 
-    await fileSystem.directory('/some/path/to/').create(recursive: true);
-    await fileSystem.file('/some/path/to/clang++').create();
-    await fileSystem.file('/some/path/to/clang').create();
-    await fileSystem.file('/some/path/to/llvm-ar').create();
-    await fileSystem.file('/some/path/to/ld.lld').create();
+    final Directory msvcBinDir = fileSystem.directory(
+        r'C:\Program Files\Microsoft Visual Studio\2022\Community\VC\Tools\MSVC\14.35.32215\bin\Hostx64\x64');
+    await msvcBinDir.create(recursive: true);
 
     final File packagesFile = fileSystem
         .directory(projectUri)
@@ -458,14 +503,20 @@ void main() {
       packagesFile,
       logger: environment.logger,
     );
-    final NativeAssetsBuildRunner runner =
-        NativeAssetsBuildRunnerImpl(projectUri, packageConfig, fileSystem, logger);
+    final NativeAssetsBuildRunner runner = NativeAssetsBuildRunnerImpl(
+      projectUri,
+      packageConfig,
+      fileSystem,
+      logger,
+    );
     final CCompilerConfig result = await runner.cCompilerConfig;
-    expect(result.cc, Uri.file('/some/path/to/clang'));
+    expect(result.cc?.toFilePath(),
+        msvcBinDir.childFile('cl.exe').uri.toFilePath());
+    expect(result.ar?.toFilePath(),
+        msvcBinDir.childFile('lib.exe').uri.toFilePath());
+    expect(result.ld?.toFilePath(),
+        msvcBinDir.childFile('link.exe').uri.toFilePath());
+    expect(result.envScript, isNotNull);
+    expect(result.envScriptArgs, isNotNull);
   });
-}
-
-class _BuildRunnerWithoutClang extends FakeNativeAssetsBuildRunner {
-  @override
-  Future<CCompilerConfig> get cCompilerConfig async => throwToolExit('Failed to find clang++ on the PATH.');
 }
